+++ src/Fluentify.Tests/AnalyzerTests.cs
@@ -1,6 +1,8 @@
 ﻿namespace Fluentify;
 
 using System.Collections.Generic;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Testing;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Testing;
@@ -13,7 +15,13 @@ public abstract class AnalyzerTests<TAnalyzer, TGenerator>
 {
     protected AnalyzerTests()
     {
+        ReferenceAssemblies = ReferenceAssemblies.Net.Net60;
         TestBehaviors = TestBehaviors.SkipGeneratedSourcesCheck;
+    }
+
+    protected override ParseOptions CreateParseOptions()
+    {
+        return new CSharpParseOptions(LanguageVersion.CSharp9);
     }
 
     protected override IEnumerable<Kind> GetSourceGenerators()