+++ src/Fluentify/AttributeAnalyzer.cs
@@ -1,41 +1,61 @@
 ﻿namespace Fluentify;
 
-using System.Collections.Immutable;
+using Fluentify.Semantics;
+using Fluentify.Syntax;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 
 /// <summary>
-/// Supports the impelmentation of analyzers relating to the specified <typeparamref name="TSyntax"/>.
+/// Supports the impelmentation of analyzers relating to <see cref="AttributeSyntax"/>.
 /// </summary>
-/// <typeparam name="TSyntax">The type of syntax upon which the analyzer will be applied.</typeparam>
-public abstract class AttributeAnalyzer<TSyntax>
-    : DiagnosticAnalyzer
-    where TSyntax : CSharpSyntaxNode
+public abstract class AttributeAnalyzer
+    : Analyzer<AttributeSyntax>
 {
-    private readonly SyntaxKind kind;
+    private readonly string name;
 
     /// <summary>
     /// Facilitates construction of an analyzer that matches for the specified <paramref name="kind"/> and
     /// may raise the specified <paramref name="diagnostics"/> if the conditions in the derived class are met.
     /// </summary>
-    /// <param name="kind">The type of syntax to match.</param>
+    /// <param name="name">The name of the attribute to match.</param>
     /// <param name="diagnostics">The rules that are applied and potentially raised by the analyzer.</param>
-    private protected AttributeAnalyzer(SyntaxKind kind, params DiagnosticDescriptor[] diagnostics)
+    private protected AttributeAnalyzer(string name, params DiagnosticDescriptor[] diagnostics)
+        : base(SyntaxKind.Attribute, diagnostics)
     {
-        this.kind = kind;
-        SupportedDiagnostics = ImmutableArray.Create(diagnostics);
+        this.name = name;
+    }
+
+    protected static bool IsViolatingMissingFluentifyRule(SyntaxNodeAnalysisContext context, AttributeSyntax syntax, out string @class, out Location location)
+    {
+        ISymbol? parent = syntax.GetParent<TypeDeclarationSyntax>(context);
+
+        if (parent is not INamedTypeSymbol type || type.HasFluentify())
+        {
+            @class = string.Empty;
+            location = Location.None;
+
+            return false;
+        }
+
+        @class = type.Name;
+        location = syntax.GetLocation();
+
+        return true;
     }
 
     /// <inheritdoc/>
-    public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; }
-
-    /// <inheritdoc/>
-    public sealed override void Initialize(AnalysisContext context)
+    protected sealed override void AnalyzeNode(SyntaxNodeAnalysisContext context, AttributeSyntax syntax)
     {
-        context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.None);
-        context.EnableConcurrentExecution();
-        context.RegisterSyntaxNodeAction(AnalyzeNode, kind);
+        IMethodSymbol? symbol = GetSymbol(context, syntax);
+
+        if (symbol is null || !IsAttribute(symbol))
+        {
+            return;
+        }
+
+        AnalyzeNode(context, symbol, syntax);
     }
 
     /// <summary>
@@ -46,33 +66,25 @@ public abstract class AttributeAnalyzer<TSyntax>
     /// Context for a syntax node action. A syntax node action can use a Microsoft.CodeAnalysis.Diagnostics.SyntaxNodeAnalysisContext
     /// to report Microsoft.CodeAnalysis.Diagnostics for a Microsoft.CodeAnalysis.SyntaxNode.
     /// </param>
+    /// <param name="symbol">
+    /// The <see cref="IMethodSymbol"/> associated with the <see cref="AttributeSyntax"/> that matches the name provided by the derivation.
+    /// </param>
     /// <param name="syntax">
     /// The <typeparamref name="TSyntax"/> matched within the <paramref name="context"/> to which the analysis is to be applied.
     /// </param>
-    protected abstract void AnalyzeNode(SyntaxNodeAnalysisContext context, TSyntax syntax);
+    protected abstract void AnalyzeNode(SyntaxNodeAnalysisContext context, IMethodSymbol symbol, AttributeSyntax syntax);
 
-    /// <summary>
-    /// Raises a <see cref="Diagnostic"/> instance within the specified <paramref name="context"/>.
-    /// </summary>
-    /// <param name="context">
-    /// Context for a syntax node action. A syntax node action can use a Microsoft.CodeAnalysis.Diagnostics.SyntaxNodeAnalysisContext
-    /// to report Microsoft.CodeAnalysis.Diagnostics for a Microsoft.CodeAnalysis.SyntaxNode.
-    /// </param>
-    /// <param name="descriptor">A <see cref="DiagnosticDescriptor"/> describing the diagnostic</param>
-    /// <param name="location">An location of the diagnostic.</param>
-    /// <param name="arguments">Arguments to the message of the diagnostic</param>
-    protected void Raise(SyntaxNodeAnalysisContext context, DiagnosticDescriptor descriptor, Location location, params object[] arguments)
+    private static IMethodSymbol? GetSymbol(SyntaxNodeAnalysisContext context, AttributeSyntax syntax)
     {
-        var diagnostic = Diagnostic.Create(descriptor, location, arguments);
-
-        context.ReportDiagnostic(diagnostic);
+        return context
+            .SemanticModel
+            .GetSymbolInfo(syntax, cancellationToken: context.CancellationToken)
+            .Symbol as IMethodSymbol;
     }
 
-    private void AnalyzeNode(SyntaxNodeAnalysisContext context)
+    private bool IsAttribute(IMethodSymbol symbol)
     {
-        if (context.Node is TSyntax syntax)
-        {
-            AnalyzeNode(context, syntax);
-        }
+        return symbol.ContainingType is not null
+            && symbol.ContainingType.IsAttribute(name);
     }
 }