+++ src/Fluentify/IgnoreAttributeAnalyzer.cs
@@ -1,50 +1,66 @@
 ﻿namespace Fluentify;
 
 using Fluentify.Semantics;
+using Fluentify.Syntax;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using static Fluentify.IgnoreAttributeAnalyzer_Resources;
+using static Fluentify.IgnoreAttributeGenerator;
 
 /// <summary>
 /// Analyzes usage of the IgnoreAttribute, ensuring the property is not already disregarded from consideration.
 /// </summary>
 [DiagnosticAnalyzer(LanguageNames.CSharp)]
 public sealed class IgnoreAttributeAnalyzer
-    : AttributeAnalyzer<PropertyDeclarationSyntax>
+    : AttributeAnalyzer
 {
     /// <summary>
     /// Facilitates construction of the analyzer.
     /// </summary>
     public IgnoreAttributeAnalyzer()
-        : base(SyntaxKind.PropertyDeclaration, RedundantUsageRule)
+        : base(Name,  MissingFluentifyRule, RedundantUsageRule)
     {
     }
 
     /// <summary>
-    /// Gets the descriptor associated with the redundant usage rule (Fluentify03).
+    /// Gets the descriptor associated with the missing fluentify rule (FLTFY05).
     /// </summary>
-    internal static DiagnosticDescriptor RedundantUsageRule { get; } = new(
-        "Fluentify03",
-        GetResourceString(nameof(Title)),
-        GetResourceString(nameof(MessageFormat)),
-        "Design",
+    internal static DiagnosticDescriptor MissingFluentifyRule { get; } = new(
+        "FLTFY05",
+        GetResourceString(nameof(MissingFluentifyRuleTitle)),
+        GetResourceString(nameof(MissingFluentifyRuleMessageFormat)),
+        "Usage",
         DiagnosticSeverity.Info,
         isEnabledByDefault: true,
-        description: GetResourceString(nameof(Description)));
+        description: GetResourceString(nameof(MissingFluentifyRuleDescription)),
+        helpLinkUri: GetHelpLinkUri("FLTFY05"));
+
+    /// <summary>
+    /// Gets the descriptor associated with the redundant usage rule (FLTFY06).
+    /// </summary>
+    internal static DiagnosticDescriptor RedundantUsageRule { get; } = new(
+        "FLTFY06",
+        GetResourceString(nameof(RedundantUsageRuleTitle)),
+        GetResourceString(nameof(RedundantUsageRuleMessageFormat)),
+        "Usage",
+        DiagnosticSeverity.Info,
+        isEnabledByDefault: true,
+        description: GetResourceString(nameof(RedundantUsageRuleDescription)),
+        helpLinkUri: GetHelpLinkUri("FLTFY06"));
 
     /// <inheritdoc/>
-    protected override void AnalyzeNode(SyntaxNodeAnalysisContext context, PropertyDeclarationSyntax syntax)
+    protected override void AnalyzeNode(SyntaxNodeAnalysisContext context, IMethodSymbol symbol, AttributeSyntax syntax)
     {
-        IPropertySymbol? property = context.SemanticModel.GetDeclaredSymbol(syntax);
-
-        if (property is null || !property.HasIgnore() || IsValidUsageOfIgnoreAttribute(property))
+        if (IsViolatingMissingFluentifyRule(context, syntax, out string @class, out Location location))
         {
-            return;
+            Raise(context, MissingFluentifyRule, location, @class);
         }
-
-        Raise(context, RedundantUsageRule, syntax.Identifier.GetLocation(), syntax.Identifier.Text);
+        else if (IsViolatingRedundantUsageRule(context, syntax, out location, out string property))
+        {
+            Raise(context, RedundantUsageRule, location, property);
+        }
     }
 
     private static LocalizableResourceString GetResourceString(string name)
@@ -52,8 +68,21 @@ public sealed class IgnoreAttributeAnalyzer
         return new(name, ResourceManager, typeof(IgnoreAttributeAnalyzer_Resources));
     }
 
-    private static bool IsValidUsageOfIgnoreAttribute(IPropertySymbol property)
+    private static bool IsViolatingRedundantUsageRule(SyntaxNodeAnalysisContext context, AttributeSyntax syntax, out Location location, out string property)
     {
-        return property.IsMutable() && property.ContainingType.HasFluentify();
+        ISymbol? symbol = syntax.GetParent(context, out PropertyDeclarationSyntax? node);
+
+        if (node is null || symbol is not IPropertySymbol declaration || declaration.IsMutable())
+        {
+            location = Location.None;
+            property = string.Empty;
+
+            return false;
+        }
+
+        location = node.GetLocation();
+        property = declaration.Name;
+
+        return true;
     }
 }